+++ client/src/main/java/com/alibaba/nacos/client/naming/net/HttpClient.java
@@ -17,7 +17,7 @@ package com.alibaba.nacos.client.naming.net;
 
 import com.alibaba.nacos.api.common.Constants;
 import com.alibaba.nacos.client.naming.utils.IoUtils;
-import com.alibaba.nacos.client.naming.utils.StringUtils;
+import com.alibaba.nacos.client.utils.StringUtils;
 import com.google.common.net.HttpHeaders;
 
 import java.io.IOException;
